backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库 - 改为 pymysql 以避免兼容性问题
pymysql==1.1.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
alembic==1.12.1

# 数据验证和序列化 - 使用兼容的版本
pydantic==2.4.2
pydantic-settings==2.0.3

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP客户端
httpx==0.25.0
aiohttp==3.8.6

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
typer==0.9.0

# 开发工具
pytest==7.4.2
pytest-asyncio==0.21.0
black==23.9.1
ruff==0.0.291